examples/otel-integration/worker/requirements.txt
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-requests==0.47b0
requests==2.31.0